Bumped version number to 0.9.6.
Added toolbars to the volume group tabs for
the same functions as the vgtree and pvtree
context menus

LvActionsMenu has been split so only the menu
itself is in it. The actions are created and enabled
in LvActions and passed to both the menu and the
new toolbar.

Created PvActionsMenu and PvActions to work like
the above for the pv context menu options and the new
toolbar.

Put a new entry into the CHANGELOG.